# docs/swagger.yaml
definitions:
...
  dto.CreateDeveloperReq:
    properties:
      api_key:
        description: API 凭证
        type: string
      login_email:
        description: 登录邮箱
        type: string
      login_pwd:
        description: 登录密码
        type: string
      name:
        description: 基础信息
        maxLength: 50
        type: string
      shared_secret:
        description: Shared Secret
        type: string
    required:
    - api_key
    - shared_secret
    type: object
  dto.CreateDraftRequest:
    properties:
      extra_prompt:
        type: string
      image_count:
        description: 1-20, 默认20
        type: integer
      quantity:
        description: 库存数量
        minimum: 1
        type: integer
      shop_ids:
        items:
          type: integer
        maxItems: 10
        minItems: 1
        type: array
      source_url:
        type: string
      style_hint:
        type: string
      user_id:
        type: integer
    required:
    - quantity
    - shop_ids
    - source_url
    - user_id
    type: object
...
  dto.UpdateDevStatusReq:
    properties:
      status:
        description: 0:Init, 1:Active, 2:Banned
        enum:
        - 0
        - 1
        - 2
        type: integer
    type: object
  dto.UpdateDeveloperReq:
    properties:
      api_key:
        description: API 凭证
        type: string
      login_email:
        description: 登录邮箱
        type: string
      login_pwd:
        description: 登录密码
        type: string
      name:
        description: 备注名称
        maxLength: 50
        type: string
      shared_secret:
        description: Shared Secret
        type: string
    type: object
...
  dto.UpdateOrderNoteRequest:
    properties:
      message_from_seller:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/developers:
    get:
      consumes:
      - application/json
      description: 分页查询开发者账号，支持按名称和状态筛选
      parameters:
      - description: 页码 (默认1)
        in: query
        name: page
        type: integer
      - description: 每页数量 (默认20)
        in: query
        name: page_size
        type: integer
      - description: 名称模糊搜索
        in: query
        name: name
        type: string
      - description: 状态 (0:未配置 1:正常 2:封禁，-1或不传表示全部)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data + total + page
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取开发者账号列表
      tags:
      - Developer (开发者账号)
    post:
      consumes:
      - application/json
      description: 录入 Etsy 开发者账号，自动生成防关联 CallbackURL
      parameters:
      - description: 创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDeveloperReq'
      produces:
      - application/json
      responses:
        "200":
          description: callback_url + 管理页地址
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建开发者账号
      tags:
      - Developer (开发者账号)
  /api/developers/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除开发者账号，同时解绑所有关联店铺
      parameters:
      - description: 开发者账号 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID 格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除开发者账号
      tags:
      - Developer (开发者账号)
    get:
      consumes:
      - application/json
      description: 根据 ID 获取单个开发者账号详细信息
      parameters:
      - description: 开发者账号 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: DeveloperResp'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID 格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取开发者账号详情
      tags:
      - Developer (开发者账号)
    put:
      consumes:
      - application/json
      description: 更新开发者账号的备注名称、登录密码或 SharedSecret
      parameters:
      - description: 开发者账号 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDeveloperReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新开发者账号信息
      tags:
      - Developer (开发者账号)
...
  /api/developers/{id}/status:
    patch:
      consumes:
      - application/json
      description: 变更开发者账号状态（启用/停用/封禁）
      parameters:
      - description: 开发者账号 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDevStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新开发者账号状态
      tags:
      - Developer (开发者账号)
  /api/drafts:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 状态筛选
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户的草稿任务列表
      tags:
      - Draft
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDraftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateDraftResult'
      summary: 提交URL创建草稿任务
      tags:
      - Draft
  /api/drafts/{task_id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DraftDetailResponse'
      summary: 获取草稿任务详情
      tags:
      - Draft
  /api/drafts/{task_id}/confirm-all:
    post:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 确认任务下所有草稿商品
      tags:
      - Draft
  /api/drafts/{task_id}/regenerate-images:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: 生成参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegenerateImagesRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 重新生成指定组的图片
      tags:
      - Draft
  /api/drafts/{task_id}/stream:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - text/event-stream
      responses: {}
      summary: SSE 实时推送任务进度
      tags:
      - Draft
...
  /api/drafts/products/{product_id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 草稿商品ID
        in: path
        name: product_id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDraftProductRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新草稿商品信息
      tags:
      - Draft
...
  /api/karrio/connections:
    get:
      description: 获取已配置的所有物流商连接
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": [], "total": 0}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取物流商连接列表
      tags:
      - Karrio (物流网关)
    post:
      consumes:
      - application/json
      description: 添加新的物流商账号连接（如 FedEx、UPS 等）
      parameters:
      - description: 连接参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{"data": {}, "message": "物流商连接创建成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建物流商连接
      tags:
      - Karrio (物流网关)
  /api/karrio/connections/{id}:
    delete:
      description: 移除指定的物流商账号连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "物流商连接已删除"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除物流商连接
      tags:
      - Karrio (物流网关)
...
  /api/karrio/rates:
    post:
      consumes:
      - application/json
      description: 根据包裹信息获取多个物流商的运费报价
      parameters:
      - description: 报价请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": []}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取运费报价
      tags:
      - Karrio (物流网关)
  /api/karrio/shipments/{id}:
    get:
      description: 获取 Karrio 系统中的运单详细信息
      parameters:
      - description: 运单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": {}}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 运单不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取 Karrio 运单详情
      tags:
      - Karrio (物流网关)
  /api/karrio/shipments/{id}/cancel:
    post:
      description: 取消 Karrio 系统中的运单
      parameters:
      - description: 运单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "运单已取消"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 取消失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 取消 Karrio 运单
      tags:
      - Karrio (物流网关)
  /api/karrio/trackers:
    get:
      description: 获取所有物流跟踪器，支持按状态筛选
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": [], "total": 0}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取跟踪器列表
      tags:
      - Karrio (物流网关)
    post:
      consumes:
      - application/json
      description: 为指定运单号创建物流跟踪
      parameters:
      - description: 跟踪参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTrackerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{"data": {}, "message": "跟踪器创建成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建物流跟踪器
      tags:
      - Karrio (物流网关)
  /api/karrio/trackers/{id}:
    get:
      description: 获取指定跟踪器的详细物流信息
      parameters:
      - description: 跟踪器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": {}}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 跟踪器不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取跟踪器详情
      tags:
      - Karrio (物流网关)
  /api/karrio/trackers/{id}/refresh:
    post:
      description: 手动刷新指定跟踪器的物流状态
      parameters:
      - description: 跟踪器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": {}, "message": "跟踪状态已刷新"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 刷新失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 刷新跟踪状态
      tags:
      - Karrio (物流网关)
  /api/karrio/trackers/batch:
    post:
      consumes:
      - application/json
      description: 批量为多个运单号创建物流跟踪
      parameters:
      - description: 批量跟踪参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchCreateTrackersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '{"data": [], "count": 0, "message": "批量创建成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 批量创建跟踪器
      tags:
      - Karrio (物流网关)
  /api/oauth/callback:
    get:
      consumes:
      - application/json
      description: 接收 Etsy 返回的 code 和 state，换取 Token 并入库
      parameters:
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 安全校验码
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 授权成功信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 拒绝授权/参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Etsy 授权回调
      tags:
      - Auth (授权模块)
  /api/oauth/login:
    get:
      consumes:
      - application/json
      description: 为店铺生成授权链接，并生成 OAuth 授权跳转链接；必传参数 region
      parameters:
      - description: 店铺 shop ID (Database Primary Key)，初次授权时填 0
        in: query
        name: shop_id
        required: true
        type: integer
      - description: 国家，必填字段
        in: query
        name: region
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 点击按钮手动复制链接 url
          schema:
            type: string
        "400":
          description: 错误信息
          schema:
            type: string
      summary: 获取 Etsy 授权链接
      tags:
      - Auth (授权模块)
  /api/orders:
    get:
      consumes:
      - application/json
      description: 分页查询订单，支持按店铺、状态、关键词、日期范围筛选
      parameters:
      - description: 店铺ID
        in: query
        name: shop_id
        type: integer
      - description: 订单状态
        in: query
        name: status
        type: string
      - description: 关键词搜索（买家名/订单号）
        in: query
        name: keyword
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.ListOrdersResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单列表
      tags:
      - Order (订单管理)
  /api/orders/{id}:
    get:
      description: 根据订单ID获取完整订单信息，包含订单项、收货地址、发货信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.OrderDetailResponse}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 订单不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单详情
      tags:
      - Order (订单管理)
  /api/orders/{id}/note:
    patch:
      consumes:
      - application/json
      description: 更新卖家对订单的备注信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 备注内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "备注已更新"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新订单备注
      tags:
      - Order (订单管理)
  /api/orders/{id}/shipment:
    get:
      description: 获取指定订单的发货记录详情
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": dto.ShipmentVO}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID格式错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 发货记录不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 发货服务未配置
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取订单发货信息
      tags:
      - Order (订单管理)
  /api/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: 更新订单的处理状态（如标记为已处理、已取消等）
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "状态已更新"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新订单状态
      tags:
      - Order (订单管理)
...
  /api/products/{id}/activate:
    post:
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 将商品状态改为 active
      tags:
      - Product
...
swagger: "2.0"